tg_client/requirements.txt
aiogram==3.26.0
aiohttp==3.13.3
python-dotenv==1.2.2
